backend/requirements.txt
chalice==1.29.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose==3.3.0
google-auth==2.23.4
python-dotenv==1.0.0
alembic==1.12.1
pydantic==2.5.2 